* master: (28 commits)
  Result of is_value_changed printed out
  unit tests: m2mreporthandler: fix uses of uninitialized memory
  unit tests: m2minterfaceimpl: fix couple of more valgrind findings
  unit tests: m2minterfaceimpl: fix a use of uninitialized memory
  unit tests: m2mresourceinstance: fix uses of uninitialized memory
  unit tests: m2mnsdlinterface: fix a bunch of uses of uninitialized memory
  version v1.8.0
  version v1.7.3
  Fix random update registration fails
  Update run_unit_tests_with_valgrind.sh
  Add script for running unit tests with valgrind
  unit tests: m2base: fix valrgind warnings for uninitialized memory
  unit tests: m2mobject: fix a use of nullptr pointed by Valgrind
  Moving TRACE_GROUP define from header to source files to avoid multiple definition error.
  Changing location
  version v1.7.2
  Change few checks in m2mtlv unit tests
  Create mbed_Client_Release_Note_16_03.md
  Creating known issues for 16.03
  Fix occasional unregistration failures
  ...

Conflicts:
	source/m2mnsdlinterface.cpp
	source/m2mobject.cpp
	source/m2mobjectinstance.cpp
	source/m2mresource.cpp
	source/m2mresourceinstance.cpp

+++ source/include/m2mnsdlinterface.h
@@ -320,9 +320,9 @@ private:
     sn_nsdl_addr_s                     _sn_nsdl_address;
     nsdl_s                            *_nsdl_handle;
     uint32_t                           _counter_for_nsdl;
-    uint16_t                           _register_id;
-    uint16_t                           _unregister_id;
-    uint16_t                           _update_id;
+    int32_t                            _register_id;
+    int32_t                            _unregister_id;
+    int32_t                            _update_id;
     uint16_t                           _bootstrap_id;
 
 friend class Test_M2MNsdlInterface;
